api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.14.0
httpx==0.27.0
python-dotenv==1.0.0
sse-starlette==1.6.5
